bring in design conventions to controller node class
