+++ src/controller/launch/controller.launch
@@ -1,0 +1,4 @@
+<launch>
+    <node name="controller_node" pkg="controller" type="controller_node" output="screen">
+    </node>
+</launch>
